A PIC18F2550 BASED PLC
This project represents Vusumuzi Tshabangu's final year project for the year 2017 in the field of elelctronic Engineering.

A PLC Emulator was designed, being an emulator meant that it didn't meet all the standards for it to be called a PLC, but is functions like a normal PLC>

The PLC was designed on a PIC18F2550 and an IDE for programming it was developed using the Netbeans IDE in Java.